+++ tb.code-discipline.ts
@@ -9,6 +9,8 @@ export default defineCodeDisciplineConfig({
     maxFileLines: { max: 350 },
     maxFunctionLines: { max: 50 },
     folderizeCompoundFiles: {},
-    dry: {},
+    dry: {
+      minDuplicateCharacters: 300,
+    },
   },
 });